EA-2420: Cleanup EntityFieldsCleanerTest
- load required configuration classes for test, instead of the entire application context
