+++ entity-management-web/src/main/java/eu/europeana/entitymanagement/config/AppConfig.java
@@ -7,7 +7,6 @@ import java.io.InputStreamReader;
 import java.util.stream.Collectors;
 
 import javax.annotation.Resource;
-import javax.validation.ValidatorFactory;
 
 import dev.morphia.Datastore;
 import eu.europeana.entitymanagement.batch.config.MongoBatchConfigurer;
@@ -17,7 +16,6 @@ import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.core.task.TaskExecutor;
-import org.springframework.validation.beanvalidation.LocalValidatorFactoryBean;
 
 import com.fasterxml.jackson.dataformat.xml.XmlMapper;
 
@@ -25,8 +23,6 @@ import eu.europeana.api.commons.oauth2.service.impl.EuropeanaClientDetailsServic
 import eu.europeana.entitymanagement.common.config.AppConfigConstants;
 import eu.europeana.entitymanagement.common.config.DataSources;
 import eu.europeana.entitymanagement.common.config.EntityManagementConfiguration;
-import eu.europeana.entitymanagement.common.config.LanguageCodes;
-import eu.europeana.entitymanagement.normalization.EntityFieldsCleaner;
 import org.springframework.web.filter.ShallowEtagHeaderFilter;
 
 /**
@@ -64,26 +60,6 @@ public class AppConfig extends AppConfigConstants{
 	}
     }
 
-    @Bean(name=BEAN_EM_LANGUAGE_CODES)
-    public LanguageCodes getLanguageCodes() throws IOException {
-
-	String languagecodesXMLConfig = emConfiguration.getLanguagecodesXMLConfig();
-	try (InputStream inputStream = getClass().getResourceAsStream(languagecodesXMLConfig);
-		BufferedReader reader = new BufferedReader(new InputStreamReader(inputStream))) {
-	    String contents = reader.lines().collect(Collectors.joining(System.lineSeparator()));
-	    return xmlMapper.readValue(contents, LanguageCodes.class);
-	}
-    }
-
-    @Bean(name=BEAN_EM_VALIDATOR_FACTORY)
-    public ValidatorFactory getValidatorFactoryBean() {
-	return new LocalValidatorFactoryBean();
-    }
-    
-    @Bean(name=BEAN_EM_ENTITY_FIELD_CLEANER)
-    public EntityFieldsCleaner getEntityFieldsCleanerBean() throws IOException {
-	return new EntityFieldsCleaner(getLanguageCodes());
-    }
     
     @Bean(name=BEAN_CLIENT_DETAILS_SERVICE)
     public EuropeanaClientDetailsService getClientDetailsService() {
