+++ entity-management-web/src/main/java/eu/europeana/entitymanagement/web/service/MetisDereferenceService.java
@@ -4,10 +4,7 @@ import eu.europeana.api.commons.error.EuropeanaApiException;
 import eu.europeana.entitymanagement.common.config.EntityManagementConfiguration;
 import eu.europeana.entitymanagement.config.AppConfig;
 import eu.europeana.entitymanagement.definitions.model.Entity;
-import eu.europeana.entitymanagement.exception.EntityCreationException;
 import eu.europeana.entitymanagement.exception.HttpBadRequestException;
-import eu.europeana.entitymanagement.web.xml.model.XmlBaseEntityImpl;
-import eu.europeana.entitymanagement.web.xml.model.metis.EnrichmentResultList;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -16,12 +13,8 @@ import org.springframework.http.MediaType;
 import org.springframework.stereotype.Service;
 import org.springframework.web.reactive.function.client.WebClient;
 
-import javax.xml.bind.JAXBContext;
-import javax.xml.bind.JAXBException;
-import javax.xml.bind.Unmarshaller;
-import java.io.StringReader;
-
 import static eu.europeana.entitymanagement.common.config.AppConfigConstants.METIS_DEREF_PATH;
+import static eu.europeana.entitymanagement.web.MetisDereferenceUtils.parseMetisResponse;
 
 /**
  * Handles de-referencing entities from Metis.
@@ -29,10 +22,6 @@ import static eu.europeana.entitymanagement.common.config.AppConfigConstants.MET
 @Service(AppConfig.BEAN_METIS_DEREF_SERVICE)
 public class MetisDereferenceService {
     private static final Logger logger = LogManager.getLogger(MetisDereferenceService.class);
-
-    private Unmarshaller jaxbDeserializer;
-    private final Object deserializerLock = new Object();
-
 
 	private final WebClient metisWebClient;
 
@@ -47,39 +36,13 @@ public class MetisDereferenceService {
      * @param id external ID for entity
      * @return An optional containing the de-referenced entity, or an empty optional
      *         if no match found.
-     * @throws EntityCreationException
+     * @throws EuropeanaApiException on error
      */
     public Entity dereferenceEntityById(String id) throws EuropeanaApiException {
 	String metisResponseBody = fetchMetisResponse(id);
 	return parseMetisResponse(id, metisResponseBody);
     }
 
-    public Entity parseMetisResponse(String id, String metisResponseBody)
-	    throws EuropeanaApiException, EntityCreationException {
-	EnrichmentResultList derefResult;
-	// Prevent "FWK005 parse may not be called while parsing" error when this method
-	// is called by multiple threads
-	synchronized (deserializerLock) {
-	    try {
-		derefResult = (EnrichmentResultList) getDeserializer().unmarshal(new StringReader(metisResponseBody));
-	    } catch (JAXBException | RuntimeException e) {
-		throw new EuropeanaApiException(
-			"Unexpected exception occurred when parsing metis dereference response for entity:  " + id, e);
-	    }
-	}
-
-	if (derefResult == null || derefResult.getEnrichmentBaseResultWrapperList().isEmpty()
-		|| derefResult.getEnrichmentBaseResultWrapperList().get(0).getEnrichmentBaseList().isEmpty()) {
-	    // Metis returns an empty XML response if de-referencing is unsuccessful,
-	    // instead of throwing an error
-	    return null;
-	}
-
-	XmlBaseEntityImpl xmlBaseEntity = derefResult.getEnrichmentBaseResultWrapperList().get(0)
-		.getEnrichmentBaseList().get(0);
-
-	return xmlBaseEntity.toEntityModel();
-    }
 
     String fetchMetisResponse(String entityId) {
 	logger.debug("De-referencing entityId={} from Metis", entityId);
@@ -100,12 +63,6 @@ public class MetisDereferenceService {
     }
 
 
-	private Unmarshaller getDeserializer() throws JAXBException {
-		if (jaxbDeserializer == null) {
-			JAXBContext jaxbContext = JAXBContext.newInstance(EnrichmentResultList.class);
-			jaxbDeserializer = jaxbContext.createUnmarshaller();
-		}
-		return jaxbDeserializer;
-	}
+
 
 }
